RALPH: Review - extract shared host config helpers
